requirements.txt
Django>=5.0,<6.0
psycopg[binary]
python-dotenv
gunicorn
whitenoise
